Make everything work in docker
