+++ src/playwright-runner.js
@@ -11,6 +11,7 @@ const { LOG_FILES } = require('./constants');
 const fs = require('fs');
 const fsExtra = require('fs-extra');
 const glob = require('glob');
+const convert = require('xml-js');
 
 const { getAbsolutePath, getArgs, exec } = utils;
 
@@ -56,6 +57,8 @@ async function createJob (suiteName, hasPassed, startTime, endTime, args, playwr
     }
   }
 
+  generateJunitfile(cwd, suiteName, args.param.browserName);
+
   let files = [
     path.join(cwd, 'console.log'),
     path.join(cwd, '__assets__', 'junit.xml'), // TOOD: Should add junit.xml.json as well
@@ -97,6 +100,79 @@ async function createJob (suiteName, hasPassed, startTime, endTime, args, playwr
   return sessionId;
 }
 
+function generateJunitfile (cwd, suiteName, browserName) {
+  let result;
+  let opts = {compact: true, spaces: 4};
+  try {
+    const xmlData = fs.readFileSync(path.join(cwd, '__assets__', `junit.xml`), 'utf8');
+    result = convert.xml2js(xmlData, opts);
+  } catch (err) {
+    console.error(err);
+  }
+  result._declaration = {
+    _attributes: {
+      version: '1.0',
+      encoding: 'utf-8'
+    }
+  };
+  result.testsuites._attributes.name = suiteName;
+  delete result.testsuites._attributes.id;
+
+  let testsuites = [];
+  let totalTests = 0;
+  let totalErrs = 0;
+  let totalFailures = 0;
+  let totalSkipped = 0;
+  let totalTime = 0;
+  for (let i = 0; i < result.testsuites.testsuite.length; i++) {
+    let testsuite = result.testsuites.testsuite[i];
+    totalTests += +testsuite._attributes.tests || 0;
+    totalFailures += +testsuite._attributes.failures || 0;
+    totalErrs += +testsuite._attributes.errors || 0;
+    totalSkipped += +testsuite._attributes.skipped || 0;
+    totalTime += +testsuite._attributes.time || 0;
+
+    testsuite._attributes.id = i;
+    let timestamp = new Date(+testsuite._attributes.timestamp);
+    testsuite._attributes.timestamp = timestamp.toISOString();
+    testsuite.properties = {};
+    testsuite.properties.property = [
+      {
+        _attributes: {
+          name: 'platformName',
+          value: process.platform,
+        }
+      },
+      {
+        _attributes: {
+          name: 'browserName',
+          value: browserName,
+        }
+      }
+    ];
+    testsuites.push(testsuite);
+  }
+  delete result.testsuites;
+  result.testsuites = {
+    _attributes: {
+      name: suiteName,
+      tests: totalTests,
+      failures: totalFailures,
+      skipped: totalSkipped,
+      errors: totalErrs,
+      time: totalTime,
+    },
+    testsuite: testsuites,
+  };
+
+  try {
+    let xmlResult = convert.js2xml(result, opts);
+    fs.writeFileSync(path.join(cwd, '__assets__', 'junit.xml'), xmlResult);
+  } catch (err) {
+    console.error(err);
+  }
+}
+
 
 async function runReporter ({ suiteName, hasPassed, startTime, endTime, args, playwright, metrics, region, metadata, saucectlVersion }) {
   let jobDetailsUrl, reportingSucceeded = false;
@@ -123,20 +199,6 @@ async function runReporter ({ suiteName, hasPassed, startTime, endTime, args, pl
 }
 
 async function run (nodeBin, runCfgPath, suiteName) {
-/*
-Options:
--c, --config <file>         Configuration file, or a test directory with optional "playwright.config.ts"/"playwright.config.js"
---forbid-only               Fail if test.only is called (default: false)
--g, --grep <grep>           Only run tests matching this regular expression (default: ".*")
---global-timeout <timeout>  Maximum time this test suite can run in milliseconds (default: unlimited)
---max-failures <N>          Stop after the first N failures
---repeat-each <N>           Run each test N times (default: 1)
---retries <retries>         Maximum retry count for flaky tests, zero for no retries (default: no retries)
---shard <shard>             Shard tests and execute only the selected shard, specify in the form "current/all", 1-based, for example "3/5"
---project <project-name>    Only run tests from the specified project (default: run all projects)
---timeout <timeout>         Specify test timeout threshold in milliseconds, zero for unlimited (default: 30000)
-*/
-
   runCfgPath = getAbsolutePath(runCfgPath);
   const runCfg = await loadRunConfig(runCfgPath);
   runCfg.path = runCfgPath;
@@ -147,17 +209,28 @@ Options:
     throw new Error(`Could not find suite named '${suiteName}'`);
   }
 
-
   const playwrightBin = path.join(__dirname, '..', 'node_modules', '.bin', 'playwright');
-  const procArgs = [playwrightBin, 'test'];
+  const procArgs = [
+    playwrightBin, 'test',
+    '--reporter', 'junit,line',
+    '--output', path.join(cwd, '__assets__'),
+    '--config', path.join(__dirname, '..', 'playwright.config.js')
+  ];
+  let args = utils.replaceLegacyKeys(suite.param);
 
-  procArgs.push('--output', path.join(cwd, '__assets__'));
-  procArgs.push('--browser', suite.param.browserName);
-  procArgs.push('--reporter', 'junit,line');
-  if (suite.param.headful) {
-    procArgs.push('--headed');
+  const excludeParams = ['screenshotOnFailure', 'video', 'slowMo'];
+
+  for (let [key, value] of Object.entries(args)) {
+    key = utils.toHyphenated(key);
+    if (excludeParams.includes(key.toLowerCase()) || value === false) {
+      continue;
+    }
+    procArgs.push(`--${key}`);
+    if (value !== true) {
+      procArgs.push(value);
+    }
   }
-  const args = _.defaultsDeep(suite, {});
+  args = _.defaultsDeep(suite, args);
 
   let env = {
     ...process.env,
