+++ tests/unit/src/playwright-runner.spec.js
@@ -5,7 +5,7 @@ jest.mock('glob');
 jest.mock('sauce-testrunner-utils');
 jest.mock('../../../src/reporter');
 const path = require('path');
-const { run } = require('../../../src/folio-runner');
+const { run } = require('../../../src/playwright-runner');
 const childProcess = require('child_process');
 const { EventEmitter } = require('events');
 const SauceLabs = require('saucelabs').default;
@@ -13,10 +13,10 @@ const fs = require('fs');
 const glob = require('glob');
 const testRunnerUtils = require('sauce-testrunner-utils');
 
-describe('folio-runner', function () {
+describe('playwright-runner', function () {
   const baseRunCfg = {
     playwright: {
-      version: '1.4.1',
+      version: '1.12.2',
       projectPath: 'path/to/project'
     },
     suites: [
@@ -32,12 +32,12 @@ describe('folio-runner', function () {
     ]
   };
   describe('.run', function () {
-    let spawnMock, folioProc, backupEnv, fsExistsMock;
+    let spawnMock, playwrightProc, backupEnv, fsExistsMock;
     beforeEach(function () {
       backupEnv = {};
       spawnMock = jest.spyOn(childProcess, 'spawn');
       fsExistsMock = jest.spyOn(fs, 'existsSync');
-      folioProc = new EventEmitter();
+      playwrightProc = new EventEmitter();
       SauceLabs.mockImplementation(() => ({
         uploadJobAssets () {
           return {};
@@ -48,9 +48,9 @@ describe('folio-runner', function () {
       }));
       spawnMock.mockImplementation(() => {
         setTimeout(() => {
-          folioProc.emit('close', 0);
+          playwrightProc.emit('close', 0);
         }, 10);
-        return folioProc;
+        return playwrightProc;
       });
       fsExistsMock.mockImplementation((url) => url.startsWith('/bad/path') ? false : true);
       process.env = {
@@ -69,7 +69,7 @@ describe('folio-runner', function () {
       procArgs[0] = path.basename(procArgs[0]);
       procArgs[procArgs.length - 1] = path.basename(procArgs[procArgs.length - 1]);
       spawnArgs.cwd = path.basename(spawnArgs.cwd);
-      spawnArgs.env.FOLIO_JUNIT_OUTPUT_NAME = path.basename(spawnArgs.env.FOLIO_JUNIT_OUTPUT_NAME);
+      spawnArgs.env.PLAYWRIGHT_JUNIT_OUTPUT_NAME = path.basename(spawnArgs.env.PLAYWRIGHT_JUNIT_OUTPUT_NAME);
       expect([nodeBin, procArgs, spawnArgs]).toMatchSnapshot();
     });
   });
