+++ tests/unit/src/utils.spec.js
@@ -23,4 +23,32 @@ describe('utils', function () {
       expect(global.logs).toMatchSnapshot();
     });
   });
+  describe('.replaceLegacyKeys', function () {
+    it('should correct vars', function () {
+      const tests = [{
+        input: {
+          browserName: 'chromium',
+        },
+        output: {
+          browser: 'chromium',
+        }
+      }, {
+        input: {
+          video: true,
+        },
+        output: {},
+      }, {
+        input: {
+          headful: true,
+        },
+        output: {
+          headed: true,
+        }
+      }];
+      for (const tt of tests) {
+        const output = utils.replaceLegacyKeys(tt.input);
+        expect(output).toMatchObject(tt.output);
+      }
+    });
+  });
 });
